Split TS library into smaller modules
